XMLHttpRequest/tests/submissions/Opera/send-accept-language.htm

<!doctype html>
<html>
  <head>
    <title>XMLHttpRequest: send() - Accept-Language</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script>
      test(function() {
        var client = new XMLHttpRequest()
        client.open("GET", "resources/accept-language.php", false)
        client.setRequestHeader("Accept-Language", "x-GameSpeak")
        client.send(null)
        assert_equals(client.responseText, "x-GameSpeak")
      })
    </script>
  </body>
</html>
